Move UserController tests in separate package. Creates test for setting balance. (#14)
