+++ src/test/kotlin/hr/foi/pop/backend/controllers/user_controller/UserControllerSetBalanceTest.kt
@@ -1,0 +1,6 @@
+package hr.foi.pop.backend.controllers.user_controller
+
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class UserControllerSetBalanceTest
